great work on io_buffer_read spec, adjusted related json parsing specs

+++ src/ion/functions/json_parse_string.h
@@ -1,3 +1,3 @@
-static inline u64 json_parse_string (
+static inline int json_parse_string (
     struct io* source
 );
